Move Illuminate\Foundation\Composer to Illuminate\Foundation\Support\Composer
