+++ src/Illuminate/Foundation/Console/AppNameCommand.php
@@ -3,9 +3,9 @@
 namespace Illuminate\Foundation\Console;
 
 use Illuminate\Console\Command;
-use Illuminate\Foundation\Composer;
 use Symfony\Component\Finder\Finder;
 use Illuminate\Filesystem\Filesystem;
+use Illuminate\Foundation\Support\Composer;
 use Symfony\Component\Console\Input\InputArgument;
 
 class AppNameCommand extends Command
@@ -27,7 +27,7 @@ class AppNameCommand extends Command
     /**
      * The Composer class instance.
      *
-     * @var \Illuminate\Foundation\Composer
+     * @var \Illuminate\Foundation\Support\Composer
      */
     protected $composer;
 
@@ -48,7 +48,7 @@ class AppNameCommand extends Command
     /**
      * Create a new key generator command.
      *
-     * @param  \Illuminate\Foundation\Composer  $composer
+     * @param  \Illuminate\Foundation\Support\Composer  $composer
      * @param  \Illuminate\Filesystem\Filesystem  $files
      * @return void
      */
