+++ src/Illuminate/Foundation/Support/Composer.php
