+++ src/Illuminate/Queue/Console/FailedTableCommand.php
@@ -4,8 +4,8 @@ namespace Illuminate\Queue\Console;
 
 use Illuminate\Support\Str;
 use Illuminate\Console\Command;
-use Illuminate\Foundation\Composer;
 use Illuminate\Filesystem\Filesystem;
+use Illuminate\Foundation\Support\Composer;
 
 class FailedTableCommand extends Command
 {
@@ -31,7 +31,7 @@ class FailedTableCommand extends Command
     protected $files;
 
     /**
-     * @var \Illuminate\Foundation\Composer
+     * @var \Illuminate\Foundation\Support\Composer
      */
     protected $composer;
 
@@ -39,7 +39,7 @@ class FailedTableCommand extends Command
      * Create a new failed queue jobs table command instance.
      *
      * @param  \Illuminate\Filesystem\Filesystem  $files
-     * @param  \Illuminate\Foundation\Composer    $composer
+     * @param  \Illuminate\Foundation\Support\Composer  $composer
      * @return void
      */
     public function __construct(Filesystem $files, Composer $composer)
